refactor: extract refresh schema table function to useRefreshSchema hooks (to be used whenever schema changes happened)
