+++ app/(frontend)/admin/schema/hooks/useSchemaAction.tsx
@@ -3,10 +3,10 @@
 import { useCallback } from "react"
 import { deleteSchema } from "@frontend/admin/schema/utils/api"
 import { toast } from "react-toastify"
-import { useSchemaTable } from "@frontend/admin/schema/hooks/useSchemaTable"
+import { useRefreshSchema } from "@frontend/admin/schema/hooks/useRefreshSchema"
 
 export const useSchemaActions = () => {
-  const { refreshSchemas } = useSchemaTable()
+  const { refreshSchemas } = useRefreshSchema()
 
   const handleDeleteSchema = useCallback(
     async (schemaId: number) => {
